fix(toggl-labs): move distraction blocker
